[Refactor] class분리
