+++ src/main/kotlin/baseball/Computer.kt
@@ -1,0 +1,18 @@
+package baseball
+
+import camp.nextstep.edu.missionutils.Randoms
+
+
+class Computer() {
+
+    fun randomNum(): MutableList<Int> {
+        val computer = mutableListOf<Int>()
+        while (computer.size < 3) {
+            val randomNumber = Randoms.pickNumberInRange(1, 9)
+            if (!computer.contains(randomNumber)) {
+                computer.add(randomNumber)
+            }
+        }
+        return computer
+    }
+}
